Split tests/test_commands.py. (#6836)
